requirements.txt
kafka-python==2.0.2
pyspark==3.4.1
elasticsearch==7.17.0
matplotlib==3.7.1
pandas==1.5.3
jupyterlab==4.0.0